requirements.txt
streamlit==1.31.1
pandas==2.2.0
google-generativeai==0.3.2
langchain>=0.1.5
langchain-google-genai==0.0.6
langchain-community>=0.0.17
chromadb==0.4.22
pdfplumber==0.10.3
python-dotenv==1.0.1
gpt4all==2.0.2
playsound
pyaudio
google-cloud-speech
scipy
numpy